+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
  - all actions use the duploctl[bot] app for commits
+ - Support for ASW S3
 
 ## [0.2.23] - 2024-04-14