+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+ - Support for AWS S3
+
 ## [0.2.25] - 2024-04-15
 
 ### Added
 
  - all actions use the duploctl[bot] app for commits
- - Support for AWS S3
+
 
 ## [0.2.23] - 2024-04-14