+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - all actions use the duploctl[bot] app for commits
 
+
 ## [0.2.23] - 2024-04-14
 
 ### Fixed